Group Parse2.hs and Command.hs.
